Adding integration tests for Tug.Server
* Using the validated Tug.Client, we can now use the client to exercise
a complementary set of tests against the Tug.Server for protocol
compatibility.
* We set this up as an integration test because we're testing the
ASP.NET Core-based Tug.Server as a whole, and not individual components,
like controllers, models or handlers.
* We make use of the ASP.NET Core TestHost which has special support for
in-process/in-memory testing from HTTP client to server.
* We include supporting artifacts to properly configure a Tug.Server
including a specially-crafted `appsettings.json` file, and supporting
assets such as MOF files and packaged modules.