Some corrections to PS5 to better handle recent changes (#49)
* Correcting for consistency with other test configs

* Updating for new parameter

An optional Agent ID has been introduced as the first parameter to the
GetModule action

* Adjusting behavior for better compliance with the expected DSC Pull Handler protocol

Some of our actions can return either a single result or null to
indicate a "Not Found" condition.  In the latter case, we were
swallowing up the condition and ended up resulting in a 500 internal
server error.