Adding a project to build a pre-configured Tug + PS5 installation
* We implement this as a PowerShell module, so that it can be easily
distributed via the PS Gallery, which seems like a natural distribution
channel for typical users of this tool.
* We build the published forms of the supporting assemblies (Tug.Server
and Ps5DscHandler) and all their dependencies, then assemble them as the
binary components embedded within the PS Module.
* We add the module manifest and number of supporting configuration
artifacts, along with PS Module cmdlets to install all the elements once
the module is installed.